superagents/orchestrator.py
"""
Agent orchestration scheduler
Coordinates when to run each superagent
"""

import asyncio
import time as time_module
from datetime import datetime
from typing import Dict, Any, List
from dotenv import load_dotenv

# Load .env file BEFORE importing config
load_dotenv()

from agents.prospector import LinkedInProspector
from agents.emailer import EmailSequencer
from agents.call_handler import CallHandler
from core.config import SuperagentConfig


class SuperagentOrchestrator:
    """Master orchestrator - schedules and runs all agents"""
    
    def __init__(self):
        self.prospector = LinkedInProspector()
        self.emailer = EmailSequencer()
        self.call_handler = CallHandler()
        self.execution_log = []
    
    async def run_daily_prospecting_schedule(self):
        """Run all daily prospecting tasks"""
        
        print("[PROSPECTING] Starting daily prospecting cycle...")
        
        results = {
            "timestamp": datetime.now().isoformat(),
            "tasks": {},
        }
        
        # 9am: Run LinkedIn prospecting
        print("[LINKEDIN] Running LinkedIn prospecting...")
        results["tasks"]["prospecting"] = await self.prospector.run_daily_prospecting()
        
        # 10am: Start email sequences for new leads
        print("[EMAIL] Scheduling email sequences...")
        # (This will happen automatically when new leads are created)
        
        # 2pm: Review email engagement
        print("[ANALYTICS] Checking email engagement...")
        # (Process opens/clicks and advance leads)
        
        # 5pm: Build call queue
        print("[OUTREACH] Preparing warm outreach...")
        
        self.execution_log.append(results)
        print(f"[SUCCESS] Daily cycle complete: {results}")
        
        return results
    
    async def run_scheduler_loop(self):
        """Main event loop - runs tasks on schedule"""
        
        print("[START] Starting Superagent Orchestrator loop...")
        print(f"[CONFIG] {SuperagentConfig.COMPANY_NAME} - {SuperagentConfig.PRODUCT_NAME}")
        
        # Run initial prospecting
        await self.run_daily_prospecting_schedule()
        
        last_prospecting = time_module.time()
        last_email_check = time_module.time()
        last_engagement_check = time_module.time()
        
        while True:
            current_time = time_module.time()
            
            # Daily prospecting (every 24 hours starting at startup)
            if current_time - last_prospecting > 86400:  # 24 hours
                print("[SCHEDULED] Running LinkedIn prospecting...")
                try:
                    await self.prospector.run_daily_prospecting()
                    last_prospecting = current_time
                except Exception as e:
                    print(f"[ERROR] Prospecting error: {e}")
            
            # Email reply check (every minute)
            if current_time - last_email_check > 60:
                try:
                    await self.emailer.process_email_replies()
                    last_email_check = current_time
                except Exception as e:
                    print(f"[ERROR] Email check error: {e}")
            
            # Engagement tracking (every 30 minutes)
            if current_time - last_engagement_check > 1800:  # 30 min
                try:
                    await self._process_email_engagement()
                    last_engagement_check = current_time
                except Exception as e:
                    print(f"[ERROR] Engagement tracking error: {e}")
            
            await asyncio.sleep(60)  # Check every minute
    
    async def _process_email_engagement(self):
        """Process email opens, clicks, replies"""
        print("[ANALYTICS] Processing email engagement metrics...")
    
    async def _run_linkedin_engagement(self):
        """Run LinkedIn engagement bot"""
        print("[LINKEDIN] Running LinkedIn engagement bot...")
    
    async def close(self):
        """Cleanup all agents"""
        await self.prospector.close()
        await self.emailer.close()
        await self.call_handler.close()


async def main():
    """Main entry point"""
    
    # Validate config
    if not SuperagentConfig.validate_required():
        print("[ERROR] Missing required configuration")
        return
    
    print("=" * 60)
    print("[FRAMEWORK] Trinity Superagent Sales Framework")
    print("=" * 60)
    SuperagentConfig.print_config()
    print("=" * 60)
    
    # Initialize orchestrator
    orchestrator = SuperagentOrchestrator()
    
    # Run main loop
    try:
        await orchestrator.run_scheduler_loop()
    except KeyboardInterrupt:
        print("\n[SHUTDOWN] Shutting down...")
        await orchestrator.close()


if __name__ == "__main__":
    asyncio.run(main())