superagents/requirements.txt
# Trinity Spine — Comprehensive Python Dependencies
# Last Updated: February 11, 2026
# Status: Production-Ready

# ==================== FastAPI & Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== Async & Concurrency ====================
asyncio==3.4.3
aiohttp==0.9.1
httpx==0.25.2

# ==================== Cryptography & Security (PQC-Enhanced) ====================
cryptography==41.0.7
liboqs-python==0.9.0  # Post-Quantum Crypto (Dilithium3 + Kyber768)
pynacl==1.5.0
# hmac: Built-in, no install needed

# ==================== Database & ORM ====================
supabase==2.3.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# ==================== API & Payments ====================
stripe==6.6.0

# ==================== Data Processing ====================
pandas==2.1.3
numpy>=1.21.0
requests>=2.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
lxml>=4.9.0
pypdf

# ==================== Monitoring & Logging ====================
prometheus-client==0.19.0
loguru==0.7.2
python-json-logger==2.0.7

# ==================== Governance & Audit ====================
dataclasses-json==0.6.1
python-dateutil==2.8.2

# ==================== Environment & Config ====================
python-dotenv==1.0.0
pyyaml==6.0.1

# ==================== Testing ====================
pytest>=7.0.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-mock==0.29.0

# ==================== Code Quality ====================
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# ==================== AI & LLM ====================
ollama>=0.0.1

# ==================== Kubernetes ====================
kubernetes==28.0.1

# ==================== Cache & Message Queue ====================
redis==5.0.1
# aio-pika==9.0.5  # Uncomment for RabbitMQ async support

# ==================== Utilities ====================
uuid6==1.0.4
